# JSON/api-docs-ms-accounts.json
{"openapi":"3.1.0","info":{"title":"API MICROSERVICE-ACCOUNTS","description":"Documentación interactiva para los servicios REST de microservice-accounts","contact":{"name":"Equipo de Desarrollo","url":"https://maveluzsoft.com","email":"soporte@maveluzsoft.com"},"license":{"name":"Licencia MIT","url":"https://opensource.org/licenses/MIT"},"version":"1.0.0"},"servers":[{"url":"http://localhost:8002","description":"Generated server url"}],"tags":[{"name":"Cuentas","description":"Operaciones relacionadas con la gestión de cuentas"},{"name":"Movimientos","description":"Operaciones relacionadas con movimientos de cuentas"}],"paths":{"/cuentas/actualizar/{id}":{"put":{"tags":["Cuentas"],"summary":"Actualizar cuenta","description":"Actualiza los datos de una cuenta existente por ID","operationId":"updateAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRegisterRequest"}}},"required":true},"responses":{"200":{"description":"Cuenta actualizada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}},"400":{"description":"Datos inválidos o cuenta nula","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}},"404":{"description":"Cuenta no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}},"500":{"description":"Error interno del servidor","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}}}}},"/movimientos/reporte":{"post":{"tags":["Movimientos"],"summary":"Obtener reporte de transacciones","description":"Genera un reporte de movimientos basado en criterios como fechas y usuario.","operationId":"getMovementsReport","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionReportRequest"}}},"required":true},"responses":{"200":{"description":"Reporte generado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}},"400":{"description":"Parámetros inválidos para el reporte"},"500":{"description":"Error al generar el reporte"}}}},"/movimientos/registrar":{"post":{"tags":["Movimientos"],"summary":"Registrar un nuevo movimiento","description":"Registra un nuevo movimiento asociado a una cuenta.","operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovementRequest"}}},"required":true},"responses":{"200":{"description":"Movimiento registrado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}},"400":{"description":"Datos de entrada inválidos"},"500":{"description":"Error al registrar el movimiento"}}}},"/cuentas/registrar":{"post":{"tags":["Cuentas"],"summary":"Registrar cuenta","description":"Registra una nueva cuenta en el sistema","operationId":"createAccount_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRegisterRequest"}}},"required":true},"responses":{"200":{"description":"Cuenta registrada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}},"400":{"description":"Datos inválidos o cuenta nula","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}},"500":{"description":"Error interno del servidor","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericAccountRegisterReponse"}}}}}}},"/movimientos/{cuentaId}":{"get":{"tags":["Movimientos"],"summary":"Listar movimientos por cuenta","description":"Obtiene la lista de movimientos asociados a una cuenta específica.","operationId":"list","parameters":[{"name":"cuentaId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Lista obtenida exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}},"400":{"description":"Solicitud incorrecta"},"500":{"description":"Error interno del servidor"}}}},"/cuentas":{"get":{"tags":["Cuentas"],"summary":"Listar cuentas","description":"Obtiene un listado de todas las cuentas registradas","operationId":"list_1","responses":{"200":{"description":"Listado obtenido correctamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericListAccountRegisterReponse"}}}}}}},"/movimientos/eliminar/{id}":{"delete":{"tags":["Movimientos"],"summary":"Eliminar un movimiento","description":"Elimina un movimiento existente por su ID.","operationId":"deleteAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Movimiento eliminado exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGeneric"}}}},"404":{"description":"Movimiento no encontrado"},"500":{"description":"Error al eliminar el movimiento"}}}},"/cuentas/eliminar/{id}":{"delete":{"tags":["Cuentas"],"summary":"Eliminar cuenta","description":"Elimina una cuenta del sistema por su ID","operationId":"deleteAccount_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"Cuenta eliminada exitosamente","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericString"}}}},"404":{"description":"Cuenta no encontrada","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericString"}}}},"500":{"description":"Error interno del servidor","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResponseGenericString"}}}}}}}},"components":{"schemas":{"AccountRegisterRequest":{"type":"object","properties":{"numeroCuenta":{"type":"string"},"tipo":{"type":"string"},"saldoInicial":{"type":"number"},"estado":{"type":"boolean"},"clienteId":{"type":"string"}}},"AccountRegisterReponse":{"type":"object","properties":{"numerCuenta":{"type":"string"},"tipo":{"type":"string"},"saldoInicial":{"type":"number"},"estado":{"type":"boolean"},"cliente":{"type":"string"}}},"ResponseGenericAccountRegisterReponse":{"type":"object","properties":{"success":{"type":"boolean"},"statusCode":{"type":"integer","format":"int32"},"timestamp":{"type":"string"},"message":{"type":"string"},"error":{"type":"string"},"data":{"$ref":"#/components/schemas/AccountRegisterReponse"},"errorCode":{"type":"string"}}},"TransactionReportRequest":{"type":"object","description":"Petición para obtener reporte por fechas y por usuario","properties":{"clientId":{"type":"string","description":"Nombre de usuario","example":"jlema"},"startDate":{"type":"string","format":"date-time","description":"Fecha y hora de inicio para el reporte (formato: yyyy-MM-dd'T'HH:mm:ss.SSS)","example":"2025-05-19T00:00:00.524"},"endDate":{"type":"string","format":"date-time","description":"Fecha y hora final para el reporte (formato: yyyy-MM-dd'T'HH:mm:ss.SSS)","example":"2025-05-19T23:59:59.524"}}},"ResponseGeneric":{"type":"object","properties":{"success":{"type":"boolean"},"statusCode":{"type":"integer","format":"int32"},"timestamp":{"type":"string"},"message":{"type":"string"},"error":{"type":"string"},"data":{"type":"object"},"errorCode":{"type":"string"}}},"MovementRequest":{"type":"object","properties":{"fecha":{"type":"string","format":"date-time"},"tipoMovimiento":{"type":"string"},"valor":{"type":"number"},"numeroCuenta":{"type":"string"}}},"ResponseGenericListAccountRegisterReponse":{"type":"object","properties":{"success":{"type":"boolean"},"statusCode":{"type":"integer","format":"int32"},"timestamp":{"type":"string"},"message":{"type":"string"},"error":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/AccountRegisterReponse"}},"errorCode":{"type":"string"}}},"ResponseGenericString":{"type":"object","properties":{"success":{"type":"boolean"},"statusCode":{"type":"integer","format":"int32"},"timestamp":{"type":"string"},"message":{"type":"string"},"error":{"type":"string"},"data":{"type":"string"},"errorCode":{"type":"string"}}}}}}